requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bidict==0.23.1
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.3
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.6.0
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.12
python-socketio==5.11.4
PyYAML==6.0.2
rich==13.9.3
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.0
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
wsproto==1.2.0
